requirements.txt
numpy
pytorch
pandas
scikit-learn
tqdm
scipy

matplotlib
tensorflow
folktables